# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Spendable Today API
  version: 1.1.0
  description: 家計データをもとに支出判断を支援するAPI
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      security: []
      summary: Login using an account provisioned in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [username, password]
              properties:
                username: { type: string, minLength: 3, maxLength: 128 }
                password: { type: string, format: password, writeOnly: true }
      responses:
        '200':
          description: A bearer session with a seven-day absolute expiry
          content:
            application/json:
              schema:
                type: object
                required: [access_token, token_type, expires_at, user]
                properties:
                  access_token: { type: string }
                  token_type: { type: string, enum: [Bearer] }
                  expires_at: { type: string, format: date-time }
                  user: { $ref: '#/components/schemas/AuthUser' }
        '400': { description: Invalid request }
        '401': { description: Invalid credentials }
        '429': { description: Rate limited. Retry after 60 seconds }
        '503': { description: Authentication database unavailable }
  /auth/me:
    get:
      summary: Return the authenticated user
      responses:
        '200':
          description: Authenticated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthUser' }
        '401': { description: Invalid or expired session }
        '503': { description: Authentication database unavailable }
  /auth/logout:
    post:
      summary: Revoke the current session
      responses:
        '204': { description: Session revoked }
        '401': { description: Invalid or expired session }
        '503': { description: Session revocation unavailable }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  schemas:
    AuthUser:
      type: object
      required: [id, username]
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
...